Added file watching
